cambios

+++ cpruebas.persistence/src/cl/eos/persistence/models/PruebaRendida.java
@@ -1,5 +1,7 @@
 package cl.eos.persistence.models;
 
+import java.text.DecimalFormat;
+
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
@@ -133,7 +135,9 @@ public class PruebaRendida implements IEntity {
 
 	public Float getPbuenas() {
 		Integer totalRespuestas = this.malas + this.buenas + this.omitidas;
-		return (float) ((this.buenas / totalRespuestas) * 100);
+//		DecimalFormat df = new DecimalFormat("#.##");
+//		System.out.print(df.format(d))
+		return (float) (((float)this.buenas / (float)totalRespuestas) * 100f);
 	}
 
 	public Integer getPuntaje() {
@@ -142,7 +146,7 @@ public class PruebaRendida implements IEntity {
 
 	public Float getPpuntaje() {
 		Integer totalRespuestas = this.malas + this.buenas + this.omitidas;
-		return (float) ((this.buenas / totalRespuestas) * 100);
+		return (float) (((float)this.buenas / (float)totalRespuestas) * 100f);
 	}
 
 }
